Squashed commit of the following:

commit bb4608db94cadaf0fa8226233d930065a0dc317b
Author: DEXREQ Automation <noreply+dexreq@oracle.com>
Date:   Fri Jul 31 19:50:25 2020 +0000

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 31 Jul 2020 19:50:25 +0000
    Running generation for: [[DEXREQ-1461, DEXREQ-1413, DEXREQ-1456, DEXREQ-1437, DEXREQ-1422, DEXREQ-1459, DEXREQ-1452, DEXREQ-1451, DEXREQ-1440, DEXREQ-1458, DEXREQ-1460, DEXREQ-1457, DEXREQ-1449]]

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 31 Jul 2020 19:23:15 +0000
    Updating pom.xml for [[DEXREQ-1461, DEXREQ-1413, DEXREQ-1456, DEXREQ-1437, DEXREQ-1422, DEXREQ-1459, DEXREQ-1452, DEXREQ-1451, DEXREQ-1440, DEXREQ-1458, DEXREQ-1460, DEXREQ-1457, DEXREQ-1449]]

    (automatically combined)

+++ operationsinsights/summarize_sql_insights_request_response.go
@@ -23,7 +23,7 @@ type SummarizeSqlInsightsRequest struct {
 	DatabaseId []string `contributesTo:"query" name:"databaseId" collectionFormat:"multi"`
 
 	// Filter sqls by percentage of db time.
-	DatabaseTimePctGreaterThan *int `mandatory:"false" contributesTo:"query" name:"databaseTimePctGreaterThan"`
+	DatabaseTimePctGreaterThan *float64 `mandatory:"false" contributesTo:"query" name:"databaseTimePctGreaterThan"`
 
 	// Specify time period in ISO 8601 format with respect to current time.
 	// Default is last 30 days represented by P30D.
